♻️ [refactor]: #452 Phase 5: lexer_skip_comment.rs を分離
skip_comment の 14 テストを `src/lexer/tests/lexer_skip_comment.rs` に分離。

- 293 件グリーン維持
- 短縮テスト関数名集合不変 / 重複 0
- lexer_skip_comment.rs: 120 行 (< 800)

Refs #452
